moved version validation to config package
